backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
groq>=0.4.2
python-dotenv>=1.0.0
SQLAlchemy>=2.0.23
cors>=1.0.1
psycopg2-binary>=2.9.9
pydantic-settings>=2.1.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.20
python-telegram-bot>=20.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
